폴더 정리 및 자잘한 오류 수정
