+++ client/src/components/cooks/CookIngreItem.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./styles/cook-ingre-item.scss";
+import "../styles/cook-ingre-item.scss";
 function CookIngreItem({ name, quan, add, active }) {
   const onAddHandler = (e) => {
     e.preventDefault();
@@ -10,7 +10,7 @@ function CookIngreItem({ name, quan, add, active }) {
     }
   };
   return (
-    <a href="#" onClick={onAddHandler} className="cook-ingre-link">
+    <a href="_blank" onClick={onAddHandler} className="cook-ingre-link">
       <div className={`cook-ingre-item ${active}`}>
         <div className="cook-ingre-name">{name}</div>
         <div className="cook-ingre-quan">{quan}</div>
