+++ client/src/components/cooks/CookPage.js
@@ -1,47 +1,41 @@
-import React, { useState } from "react";
-import "./styles/mypage.scss";
-import { FaQuoteLeft, FaQuoteRight, FaCog } from "react-icons/fa";
+import React from "react";
+import qs from "qs";
 import { useEffect } from "react";
-import TimeLine from "./TimeLine";
+import { useState } from "react";
+import "../styles/postsStyle/tag-page.scss";
+import TimeLine from "../TimeLine";
 import Axios from "axios";
-import { debounce } from "lodash";
-import { useCallback } from "react";
 import { useRef } from "react";
-import SkeletonLoading from "./loadingCompo/SkeletonLoading";
-import SkeletonMypage from "./loadingCompo/SkeletonMypage";
-import NotFound from "./NotFound";
-
-function ChefPage({ match }) {
-  //다른 유저들의 페이지
-  const [sort, setSort] = useState("mypost");
-  const [postList, setList] = useState([]);
-  const [userInfo, setUserInfo] = useState({
-    user_nickname: "",
-    user_description: "",
-    user_img: "",
+import { useCallback } from "react";
+import { debounce } from "lodash";
+import SkeletonLoading from "../loadingCompo/SkeletonLoading";
+import NotFound from "../NotFound";
+function CookPage({ location }) {
+  const ingres = qs.parse(location.search, {
+    ignoreQueryPrefix: true, //?뺴고 파싱
   });
+  const [sort, setSort] = useState("exact");
+  const [postList, setList] = useState([]);
   const [loading, setLoading] = useState(false);
-  const [userLoading, setUserLoading] = useState(false);
   const [isEnd, setIsEnd] = useState(false); //더 이상 가져올 수 있는 데이터가 있는지
   const [isFirst, setisFirst] = useState(true);
-  const [prevName, setPrevName] = useState(""); //라우트로 이동시 전페이지의 값과 비교하기 위해 사용
+  const [prevName, setPrevName] = useState("");
+
   const limitItem = useRef(0); //페이징 처리용 리미트
   const isMounted = useRef(null);
-
-  const { user_nickname, user_description, user_img } = userInfo;
   const sortList = [
     //정렬 리스트
-    { name: "mypost", koName: "레시피" },
-    { name: "scrap", koName: "스크랩" },
+    { name: "exact", koName: "정확순" },
+    { name: "latest", koName: "최신순" },
+    { name: "popular", koName: "인기순" },
   ];
   const onSortChange = (target) => () => {
     if (loading) {
       setSort(target);
-      if (user_nickname) {
-        getPosts(target);
-      }
+      getPosts(target);
     }
   };
+
   const sortItem = sortList.map((item, index) => (
     //만들어논 리스트를 맵으로 돌린후 저장한다.
     //이 과정에서 item의 이름이 현재 상태와 같으면 active를 클래스로 준다.
@@ -55,65 +49,6 @@ function ChefPage({ match }) {
     </button>
   ));
 
-  const getUser = (params = match.params.no) => {
-    setUserLoading(false);
-    Axios.get(`/api/users/profile/${params}`)
-      .then((res) => {
-        if (isMounted.current) {
-          if (res.data.success) {
-            setUserInfo(res.data.result);
-            setUserLoading(true);
-          } else {
-            alert(res.data.message);
-          }
-        }
-      })
-      .catch((err) => console.log(err));
-  };
-  const userData = () => {
-    //렌더링 될 포스트 데이터
-    if (userLoading) {
-      if (user_nickname) {
-        return (
-          <div className="my-top">
-            <div className="my-top-description-box">
-              <div className="description-text">
-                <FaQuoteLeft className="quote-left" />
-                <span>
-                  {user_description
-                    ? user_description
-                    : "등록된 자기소개가 없습니다."}
-                </span>
-                <FaQuoteRight className="quote-right" />
-              </div>
-            </div>
-            <div className="my-top-sort-box">
-              <div className="my-img-box">
-                {user_img ? (
-                  <img
-                    src={`http://localhost:5000/${user_img}`}
-                    alt="프로필사진"
-                  />
-                ) : (
-                  <img
-                    src={`http://localhost:5000/uploads/normal-profile.png`}
-                    alt="프로필사진"
-                  />
-                )}
-              </div>
-              {sortItem}
-              <div className="user-name">{user_nickname}</div>
-            </div>
-          </div>
-        );
-      } else {
-        //로딩이 다되었음에도 불구하고 length 가 0이면 결과가 없는것이다.
-        return <NotFound item={"유저"} />;
-      }
-    } else {
-      return null;
-    }
-  };
   const postsData = () => {
     //렌더링 될 포스트 데이터
     if (loading) {
@@ -137,17 +72,21 @@ function ChefPage({ match }) {
         ));
       } else {
         //로딩이 다되었음에도 불구하고 length 가 0이면 결과가 없는것이다.
-        return null;
+        return <NotFound item={"재료"} />;
       }
     } else {
-      return "loading...";
+      return null;
     }
   };
-  const getPosts = (target, params = match.params.no) => {
+
+  const getPosts = (target, params = ingres.names) => {
+    window.scrollTo(0, 0);
     limitItem.current = 0; //0으로 초기화
     setLoading(false);
     setIsEnd(false);
-    Axios.get(`/api/users/posts/${params}/${target}/${limitItem.current}`)
+    Axios.get(
+      `/api/post/cook/getposts/${params}/${target}/${limitItem.current}`
+    )
       .then((res) => {
         if (isMounted.current) {
           if (res.data.success) {
@@ -156,7 +95,9 @@ function ChefPage({ match }) {
               console.log("isEnd");
               setIsEnd(true); //10개씩 가져오는데 그것 보다 작으면 그것이 최대이다.
             } else {
-              limitItem.current += 10; //10개보다 크거나 같을때만 +10을 해준다.
+              if (limitItem.current === 0) {
+                limitItem.current += 10; //10개보다 크거나 같을때만 +10을 해준다.
+              }
             }
             setLoading(true);
           } else {
@@ -165,18 +106,23 @@ function ChefPage({ match }) {
         }
       })
       .catch((err) => console.log(err));
+    console.log(limitItem.current);
   };
-
   const loadPosts = (target) => {
     console.log(target);
+    console.log(postList);
+    console.log(ingres.names);
     if (!isEnd) {
       console.log(isEnd);
+      console.log(limitItem.current);
       Axios.get(
-        `/api/users/posts/${match.params.no}/${target}/${limitItem.current}`
+        `/api/post/cook/getposts/${ingres.names}/${target}/${limitItem.current}`
       )
         .then((res) => {
           if (isMounted.current) {
             if (res.data.success) {
+              console.log("결과");
+              console.log(postList);
               setList([...postList, ...res.data.result]);
               if (res.data.result.length < 10) {
                 console.log("isEnd");
@@ -198,14 +144,17 @@ function ChefPage({ match }) {
   const onChangeRoute = useCallback(
     debounce(
       (params) => {
+        console.log("dd");
         //라우트가 과도하게 변경될때에 대한 디바운스처리로, match를 deps에 넣으면 무조건 바뀌므로 인자로 받아온다.
-        if (prevName && prevName !== params) {
-          window.scrollTo(0, 0);
+        if (prevName && prevName !== params && isMounted.current) {
           setPrevName(params); //요청이 먼저 가버리면 다시 실행되기 때문에 가장 먼저 변경
           console.log("성공");
-          getPosts("mypost", params);
-          getUser(params);
-          setSort("mypost");
+          console.log(loading);
+          console.log(postList);
+          console.log(prevName);
+          console.log(params);
+          getPosts("exact", params);
+          setSort("exact");
         }
       },
       300,
@@ -213,7 +162,6 @@ function ChefPage({ match }) {
     ),
     [prevName]
   );
-
   const onScrollHandler = useCallback(() => {
     const { clientHeight } = document.documentElement; //요소 높이
     const scrollTop = //남은 높이
@@ -228,44 +176,42 @@ function ChefPage({ match }) {
       loading
     ) {
       console.log("맨끝");
-      console.log(scrollTop);
-      console.log(clientHeight);
-      console.log(scrollHeight);
       loadPosts(sort);
     }
-
-    //deps를 넣어줘야 최신 상태를 유지할 수 있다.
+    //deps를 넣어줘야 최신 상태를 유지할 수 있다. 이벤트리스너에 등록하는경우 props나 state를 따르지 않는다.
   }, [postList, isEnd, sort, loading]);
 
   useEffect(() => {
     isMounted.current = true;
-    console.log("asdsad");
-    console.log(match.params.no);
+    console.log("??");
+    console.log(loading);
     if (isFirst) {
-      window.scrollTo(0, 0);
-      getPosts("mypost");
-      getUser();
+      console.log("처음");
+      getPosts("exact");
       setisFirst(false);
-      setPrevName(match.params.no);
+      setPrevName(ingres.names);
     }
-    onChangeRoute(match.params.no);
+    onChangeRoute(ingres.names);
     window.addEventListener("scroll", onScrollHandler);
     return () => {
       window.removeEventListener("scroll", onScrollHandler);
       isMounted.current = false;
     };
-  }, [onScrollHandler, match.params.no, user_nickname]);
+  }, [onScrollHandler, ingres.names]);
 
   return (
-    <>
-      <div className="mypage-container">
-        <SkeletonMypage active={userLoading} />
-        {userData()}
-        <SkeletonLoading limit={5} active={loading} />
-        {postsData()}
+    <div className="tag-page-container">
+      <div className="tag-page-title">
+        <span className="tag-page-title-tag">{`"${ingres.names}"`}</span>
+        <span>에 대한 검색결과가 </span>
+        <span className="tag-page-title-index">{postList.length}</span>
+        <span>개 있습니다.</span>
       </div>
-    </>
+      <div className="sort-box">{sortItem}</div>
+      <SkeletonLoading limit={5} active={loading} />
+      {postsData()}
+    </div>
   );
 }
 
-export default ChefPage;
+export default CookPage;
