+++ client/src/components/styles/cook-modal.scss
@@ -89,6 +89,10 @@
         position: relative;
         display: flex;
         align-items: center;
+        form {
+          width: 100%;
+          position: relative;
+        }
         input {
           width: 100%;
           height: 40px;
@@ -103,6 +107,7 @@
           position: absolute;
           width: 18%;
           height: 40px;
+          top: 0;
           left: 100%;
           transform: translateX(-100%);
           border: none;
