+++ client/src/components/styles/postsStyle/post-header.scss
@@ -18,6 +18,10 @@
       span {
         margin-bottom: 10px; //공통
       }
+      .loading-spinner {
+        align-self: flex-end;
+        margin-right: 10px;
+      }
       .edit-btn-box {
         align-self: flex-end;
         button {
