+++ client/src/reducer/index.js
@@ -1,5 +1,4 @@
 import { combineReducers } from "redux";
-import storage from "redux-persist/lib/storage";
 import user from "../modules/user";
 import write from "../modules/write";
 import post from "../modules/post";
